feat: chỉnh sửa lại giao diện ứng dụng - loại bỏ icon add ở bottom navigation -> di chuyển nó đến màn home - loại bỏ tab explore - thêm màn AI Generative (tiền thân là màn explore) - Chỉnh sửa lỗi tạo study set bị crash (cho vào trong try-catch)
