+++ app/src/main/java/com/pwhs/quickmem/presentation/BottomNavItem.kt
@@ -3,7 +3,6 @@ package com.pwhs.quickmem.presentation
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import com.pwhs.quickmem.R
-import com.ramcosta.composedestinations.generated.destinations.ExploreScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.HomeScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.LibraryScreenDestination
 import com.ramcosta.composedestinations.generated.destinations.ProfileScreenDestination
@@ -19,20 +18,6 @@ sealed class BottomNavItem(
         title = R.string.txt_home,
         icon = R.drawable.ic_home,
         route = HomeScreenDestination.route,
-        direction = HomeScreenDestination()
-    )
-
-    data object Explore : BottomNavItem(
-        title = R.string.txt_explore,
-        icon = R.drawable.ic_explore,
-        route = ExploreScreenDestination.route,
-        direction = ExploreScreenDestination()
-    )
-
-    data object Center : BottomNavItem(
-        title = R.string.txt_center_fab,
-        icon = R.drawable.ic_add_circle,
-        route = "fab",
         direction = HomeScreenDestination()
     )
