--- app/src/main/java/com/pwhs/quickmem/presentation/app/explore/ExploreTabEnum.kt
@@ -1,6 +1,0 @@
-package com.pwhs.quickmem.presentation.app.explore
-
-enum class ExploreTabEnum(val index: Int) {
-    TOP_STREAK(1),
-    CREATE_STUDY_SET(0)
-}
