--- app/src/main/java/com/pwhs/quickmem/presentation/app/explore/ExploreUiEvent.kt
@@ -1,9 +1,0 @@
-package com.pwhs.quickmem.presentation.app.explore
-
-import androidx.annotation.StringRes
-
-sealed class ExploreUiEvent {
-    data class Error(@StringRes val message: Int) : ExploreUiEvent()
-    data class EarnedCoins(@StringRes val message: Int) : ExploreUiEvent()
-    data class CreatedStudySet(val studySetId: String) : ExploreUiEvent()
-}
