+++ app/src/main/java/com/pwhs/quickmem/presentation/app/library/LibraryScreen.kt
@@ -163,10 +163,10 @@ fun LibraryScreen(
             navigator.navigate(CreateStudySetScreenDestination())
         },
         navigateToCreateClass = {
-            navigator.navigate(CreateClassScreenDestination)
+            navigator.navigate(CreateClassScreenDestination())
         },
         navigateToCreateFolder = {
-            navigator.navigate(CreateFolderScreenDestination)
+            navigator.navigate(CreateFolderScreenDestination())
         }
     )
 }
