+++ app/src/main/res/values-vi/strings.xml
@@ -678,6 +678,7 @@
     <string name="txt_no_data">Chưa có dữ liệu</string>
     <string name="txt_statistics">Thống kê</string>
     <string name="txt_failed_to_clear_all_notifications">Lỗi khi xoá tất cả thông báo</string>
+    <string name="txt_ai_generative"> AI sinh sáng</string>
     <plurals name="hours">
         <item quantity="other">%d giờ</item>
     </plurals>
