+++ app/src/main/res/values/strings.xml
@@ -682,6 +682,7 @@
     <string name="txt_no_data">No data</string>
     <string name="txt_statistics">Statistics</string>
     <string name="txt_failed_to_clear_all_notifications">Fail to clear all notifications</string>
+    <string name="txt_ai_generative">AI Generative</string>
     <plurals name="hours">
         <item quantity="one">%d hour</item>
         <item quantity="other">%d hours</item>
